reorganize for legacy
