+++ archived/TextHighligher.ts
