+++ archived/TextHighlighter.js
