+++ lib/Utils.js
@@ -1,31 +1,21 @@
 "use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.paramsImp = exports.optionsImpl = exports.activator = exports.groupHighlights = exports.defaults = exports.highlightI = exports.haveSameColor = exports.unique = exports.sortByDepth = exports.refineRangeBoundaries = exports.dom = exports.IGNORE_TAGS = exports.NODE_TYPE = exports.TIMESTAMP_ATTR = exports.DATA_ATTR = void 0;
-// eslint-disable-next-line @typescript-eslint/class-name-casing
-var highlightI = /** @class */ (function () {
-    function highlightI() {
-    }
-    return highlightI;
-}());
-exports.highlightI = highlightI;
-// eslint-disable-next-line @typescript-eslint/class-name-casing
-var optionsImpl = /** @class */ (function () {
-    function optionsImpl() {
-    }
-    return optionsImpl;
-}());
-exports.optionsImpl = optionsImpl;
-// class containerI{
-//     querySelectorAll: (arg0: string): any;
-//     hasAttribute: (arg0: string) => any;
+// highlight extensions
+// eslint-disable-next-line @typescript-eslint/class-name-casing,@typescript-eslint/camelcase
+// type H_HTMLElement = HTMLElement;
+// eslint-disable-next-line @typescript-eslint/class-name-casing,@typescript-eslint/camelcase
+// interface ie_HTMLElement extends HTMLElement {
+//   createTextRange(): TextRange;
 // }
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.activator = exports.groupHighlights = exports.defaults = exports.haveSameColor = exports.unique = exports.sortByDepth = exports.refineRangeBoundaries = exports.dom = exports.IGNORE_TAGS = exports.NODE_TYPE = exports.TIMESTAMP_ATTR = exports.DATA_ATTR = void 0;
 // eslint-disable-next-line @typescript-eslint/class-name-casing
-var paramsImp = /** @class */ (function () {
-    function paramsImp() {
-    }
-    return paramsImp;
-}());
-exports.paramsImp = paramsImp;
+// interface H_Node extends Node {
+//   splitText(endOffset: number): any;
+// }
+// eslint-disable-next-line @typescript-eslint/class-name-casing,@typescript-eslint/camelcase
+// interface H_Window extends Window {
+//   find(text: any, caseSens: any): boolean;
+// }
 var /**
    * Attribute added by default to every highlight.
    * @type {string}
